requirements.txt
streamlit
pandas
sentence-transformers
scikit-learn
google-generativeai
python-docx
PyPDF2
nltk
